1.73.0 (5.3.8p2)
Build-Id: 9116e27c-e686-474b-891e-17a93574ddcc
Build-Sha256: 7aab335d434ca3fdd96dcf1812f1d1ffe4e3ebdc5ce89239c5c50d74a28a0e01

+++ Assembly-CSharp/DeathUIController.cs
@@ -23,7 +23,7 @@ public class DeathUIController : MonoBehaviour
 
 	private void OnAvatarStateChanged(object state)
 	{
-		AvatarModeTypes avatarModeTypes = (AvatarModeTypes)state;
+		AvatarModeTypes avatarModeTypes = (AvatarModeTypes)(int)state;
 		if ((avatarModeTypes & AvatarModeTypes.Hidden) != 0)
 		{
 			fader.Deactivate();
