+++ Assembly-CSharp/WaterPlaneManager.cs
@@ -183,7 +183,7 @@ public class WaterPlaneManager : MonoBehaviour
 			avatarInteractable.AddModifier(AvatarModifierPackageType.Underwater, -1, additionalUnderWaterModifiers);
 			if (waterPlaneLogicCube != null && waterPlaneLogicCube.Data.ContainsKey("avatarModifierPackageType"))
 			{
-				AvatarModifierPackageType avatarModifierPackageType = (AvatarModifierPackageType)waterPlaneLogicCube.Data["avatarModifierPackageType"];
+				AvatarModifierPackageType avatarModifierPackageType = (AvatarModifierPackageType)(int)waterPlaneLogicCube.Data["avatarModifierPackageType"];
 				if (avatarModifierPackageType != AvatarModifierPackageType.None)
 				{
 					AvatarModifierPackage package = AvatarModifierPackageFactory.GetPackage(avatarModifierPackageType);
